Move AnalogStick.tsx  + extract RCSelectMenu from RC panel component
